Prevent Validation tab from being enabled when no question is selected.
Added an AbstractFormDesignerView class. Currently it has one method, setAllEnabled() which takes an instance of HasWidgets and a boolean value (for enabled/disabled) and recursively enables/disables all child widgets. Need to update the implementation of this method to use the GWT HasEnabled interface, but need to get the project to build under GWT 2.1.0 first. This is a temporary fix.

Changed the setEnabled() method in ValidationRulesView to call setAllEnabled() on its root panel for correct enable/disable behavior, then added the appropriate calls to setEnabled in the setQuestionDef() method.

Changed GroupOperationWidget to implement an interim HasEnabled interface of my own making so that it will behave correctly with enable/disable until I get GWT 2.1.0 integrated. 

Changed the "Add New Condition" button in SkipRulesView to a "Add Condition" button.

git-svn-id: https://trac.openxdata.org/svn/openxdata/Forms/branches/usm-project@3493 6cddc80d-1c40-44cd-927a-2be6005847d2